fix file split
